+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added support for virtualenv autogeneration (Python 3.6+ only).
 - Added the `run` command to execute commands inside the created virtualenvs.
 - Added the `debug:resolve` command to debug dependency resolution.
+- Added `pyproject.toml` file validation.
 
 ### Fixed